+++ src/corrente/attributes/styles.ts
@@ -6,7 +6,7 @@ import { coerceToObservable, NullLike, StringLike, TypeOrObservable } from "../u
 import { AttributeMetadataDictionary, AttributeMetadataObject, setAttributes } from "./attribute-operator-isolation";
 
 /**
- * The style names which may be applied to an RxFM element.
+ * The style names which may be applied to an Corrente element.
  */
 export type StyleKeys = Extract<
   keyof { [K in keyof CSSStyleDeclaration as CSSStyleDeclaration[K] extends string ? K : never]: CSSStyleDeclaration[K] },
@@ -106,7 +106,7 @@ type StyleOperators = {
 export type StyleOperator = BasicStyleOperator & StyleOperators;
 
 /**
- * An observable operator to manage a style on an RxFM component.
+ * An observable operator to manage a style on an Corrente component.
  * Alternatively style operators for specific style types may be accessed directly as properties eg: `style.color('red')`.
  * @param name The style name.
  * @param value The style value or an observable emitting the value.
@@ -121,7 +121,7 @@ export type Styles = {
 };
 
 /**
- * An observable operator to update the styles on an RxFM component.
+ * An observable operator to update the styles on an Corrente component.
  * @param stylesOrObservableStyles A dictionary (or observable emitting a dictionary) of style names to values.
  * Values may be strings, null-like or observables emitting these.
  */